Extract a BibRecord mixin
